Collate "_bin" functionality into collections.
